Create useReadComments hook
