+++ src/lib/hooks/index.ts
@@ -8,6 +8,7 @@ export { default as useUpdatePost } from './useUpdatePost';
 
 export { default as useCreateComment } from './useCreateComment';
 export { default as useDeleteComment } from './useDeleteComment';
+export { default as useReadComments } from './useReadComments';
 export { default as useUpdateComment } from './useUpdateComment';
 
 export { default as useMutateReaction } from './useMutateReaction';
